+++ CHANGELOG.md
@@ -5,7 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.0.1] - 2020-05-27
+### Added
+- Add dependency: "phlak/semver"
+
+### Removed
+- Dependencies for "graham-campbell/markdown"
+
+## [1.0.0] - 2020-05-27
 ### Added
 - Basic support for changelog creation / line updated.
-- Support for custom provider
+- Basic support for release.
+- Support for custom configuration.
+- Unit tests for ChangelogHelper.